refactor(LineRepository) : deleteAll 쿼리문 수정
 "TRUNCATE TABLE LINE" -> "DELETE FROM line"
