+++ src/main/java/wooteco/subway/line/repository/infra/JDBCLineRepository.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line.repository;
+package wooteco.subway.line.repository.infra;
 
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -8,6 +8,7 @@ import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import wooteco.subway.exception.DuplicatedNameException;
 import wooteco.subway.line.domain.Line;
+import wooteco.subway.line.repository.LineRepository;
 
 import java.sql.PreparedStatement;
 import java.util.ArrayList;
@@ -32,7 +33,7 @@ public class JDBCLineRepository implements LineRepository {
     @Override
     public Line save(final Line line) {
         try {
-            String query = "INSERT INTO LINE(name, color) VALUES(?, ?)";
+            String query = "INSERT INTO line(name, color) VALUES(?, ?)";
             KeyHolder keyHolder = new GeneratedKeyHolder();
             this.jdbcTemplate.update(connection -> {
                 PreparedStatement ps = connection.prepareStatement(query, new String[]{"id"});
@@ -48,31 +49,31 @@ public class JDBCLineRepository implements LineRepository {
 
     @Override
     public List<Line> findAll() {
-        String query = "SELECT * FROM LINE";
+        String query = "SELECT * FROM line";
         return this.jdbcTemplate.query(query, actorRowMapper);
     }
 
     @Override
     public Line findById(final Long id) {
-        String query = "SELECT * FROM LINE WHERE id = ?";
+        String query = "SELECT * FROM line WHERE id = ?";
         return this.jdbcTemplate.queryForObject(query, actorRowMapper, id);
     }
 
     @Override
     public void delete(final Long id) {
-        String query = "DELETE FROM LINE WHERE id = ?";
+        String query = "DELETE FROM line WHERE id = ?";
         this.jdbcTemplate.update(query, id);
     }
 
     @Override
     public void update(final Line line) {
-        String query = "UPDATE LINE SET name = ?, color = ? WHERE id = ?";
+        String query = "UPDATE line SET name = ?, color = ? WHERE id = ?";
         this.jdbcTemplate.update(query, line.getName(), line.getColor(), line.getId());
     }
 
     @Override
     public void deleteAll() {
-        String query = "TRUNCATE TABLE LINE";
+        String query = "DELETE FROM line";
         this.jdbcTemplate.update(query);
     }
 }
